(HP-1657) Split ServiceConnections to own ServiceConnection component
